Make libscript optional

+++ m
@@ -8,23 +8,24 @@ PROGRAM=dit
 make=make
 [ -e /bin/ColorMake ] && make=ColorMake
 
-if [ ! -e mode.txt -o "`cat mode.txt`" != "$1" ]
+if [ -e $PROGRAM ] && [ ! -e mode.txt -o "`cat mode.txt`" != "$1" ]
 then
    cp $PROGRAM $PROGRAM.old
    make clean
    mv $PROGRAM.old $PROGRAM
 fi
-echo "$1" > mode.txt
+mode=`echo "$@" | tr ' ' '\n' | grep -v -- --`
+echo $mode > mode.txt
 
 HEADERS=(Prototypes.h Structures.h)
 for h in "${HEADERS[@]}"
 do
-   mv $h $h.old
+   [ -e $h ] && mv $h $h.old
 done
 ./GenHeaders
 for h in "${HEADERS[@]}"
 do
    diff --text $h $h.old &>/dev/null && mv $h.old $h
 done
-./MakeMakefile $PROGRAM > Makefile
-$make $1
+./MakeMakefile $PROGRAM "$@" > Makefile
+$make $mode
